streamlit-app/requirements.txt
streamlit
chromadb
anthropic
sentence-transformers
python-dotenv
